Move token controller to a iiif/auth/v1 directory
This gives us a place to add the v2 auth
