+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2023-11-08 16:02:12 UTC using RuboCop version 1.57.2.
+# on 2023-12-06 16:20:31 UTC using RuboCop version 1.58.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -15,7 +15,7 @@ Layout/MultilineMethodCallBraceLayout:
     - 'spec/routing/file_routing_spec.rb'
     - 'spec/routing/media_routing_spec.rb'
 
-# Offense count: 7
+# Offense count: 8
 # Configuration parameters: AllowedMethods, AllowedPatterns, CountRepeatedAttributes.
 Metrics/AbcSize:
   Max: 22
@@ -25,12 +25,12 @@ Metrics/AbcSize:
 Metrics/CyclomaticComplexity:
   Max: 10
 
-# Offense count: 12
+# Offense count: 13
 # Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 25
 
-# Offense count: 31
+# Offense count: 25
 RSpec/AnyInstance:
   Exclude:
     - 'spec/controllers/media_controller_spec.rb'
@@ -39,19 +39,18 @@ RSpec/AnyInstance:
     - 'spec/features/status_spec.rb'
     - 'spec/requests/file_auth_request_spec.rb'
     - 'spec/requests/file_spec.rb'
+    - 'spec/requests/iiif/auth/v1/tokens_spec.rb'
     - 'spec/requests/iiif_auth_request_spec.rb'
     - 'spec/requests/iiif_spec.rb'
     - 'spec/requests/media_auth_request_spec.rb'
 
-# Offense count: 78
+# Offense count: 68
 # Configuration parameters: Prefixes, AllowedPatterns.
 # Prefixes: when, with, without
 RSpec/ContextWording:
   Exclude:
     - 'spec/abilities/ability_spec.rb'
     - 'spec/controllers/file_controller_spec.rb'
-    - 'spec/controllers/iiif_controller_spec.rb'
-    - 'spec/controllers/iiif_token_controller_spec.rb'
     - 'spec/controllers/legacy_image_service_controller_spec.rb'
     - 'spec/controllers/media_controller_spec.rb'
     - 'spec/features/status_spec.rb'
@@ -72,7 +71,7 @@ RSpec/DescribedClass:
     - 'spec/models/stacks_media_token_spec.rb'
     - 'spec/models/user_spec.rb'
 
-# Offense count: 7
+# Offense count: 4
 # This cop supports safe autocorrection (--autocorrect).
 RSpec/EmptyLineAfterExampleGroup:
   Exclude:
@@ -80,7 +79,7 @@ RSpec/EmptyLineAfterExampleGroup:
     - 'spec/controllers/legacy_image_service_controller_spec.rb'
     - 'spec/requests/file_auth_request_spec.rb'
 
-# Offense count: 32
+# Offense count: 34
 # This cop supports safe autocorrection (--autocorrect).
 RSpec/EmptyLineAfterFinalLet:
   Exclude:
@@ -94,7 +93,7 @@ RSpec/EmptyLineAfterFinalLet:
     - 'spec/services/iiif_metadata_service_spec.rb'
     - 'spec/services/media_authentication_json_spec.rb'
 
-# Offense count: 6
+# Offense count: 7
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: AllowConsecutiveOneLiners.
 RSpec/EmptyLineAfterHook:
@@ -103,24 +102,23 @@ RSpec/EmptyLineAfterHook:
     - 'spec/features/status_spec.rb'
     - 'spec/requests/iiif_spec.rb'
 
-# Offense count: 12
+# Offense count: 11
 # This cop supports safe autocorrection (--autocorrect).
 RSpec/EmptyLineAfterSubject:
   Exclude:
     - 'spec/abilities/ability_spec.rb'
-    - 'spec/controllers/iiif_controller_spec.rb'
     - 'spec/controllers/webauth_controller_spec.rb'
     - 'spec/models/iiif_image_spec.rb'
     - 'spec/models/projection_spec.rb'
     - 'spec/models/stacks_image_spec.rb'
     - 'spec/services/iiif_metadata_service_spec.rb'
 
-# Offense count: 20
+# Offense count: 22
 # Configuration parameters: CountAsOne.
 RSpec/ExampleLength:
-  Max: 24
+  Max: 16
 
-# Offense count: 15
+# Offense count: 12
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: CustomTransform, IgnoredWords, DisallowedExamples.
 # DisallowedExamples: works
@@ -131,7 +129,7 @@ RSpec/ExampleWording:
     - 'spec/models/stacks_media_token_spec.rb'
     - 'spec/requests/iiif_auth_request_spec.rb'
 
-# Offense count: 3
+# Offense count: 2
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: implicit, each, example
@@ -140,13 +138,11 @@ RSpec/HookArgument:
     - 'spec/requests/file_auth_request_spec.rb'
     - 'spec/requests/iiif_auth_request_spec.rb'
 
-# Offense count: 8
+# Offense count: 6
 # This cop supports safe autocorrection (--autocorrect).
 RSpec/LeadingSubject:
   Exclude:
     - 'spec/controllers/file_controller_spec.rb'
-    - 'spec/controllers/iiif_controller_spec.rb'
-    - 'spec/controllers/iiif_token_controller_spec.rb'
     - 'spec/models/projection_spec.rb'
     - 'spec/models/stacks_media_token_spec.rb'
     - 'spec/services/media_authentication_json_spec.rb'
@@ -159,19 +155,17 @@ RSpec/MessageSpies:
     - 'spec/controllers/file_controller_spec.rb'
     - 'spec/controllers/media_controller_spec.rb'
 
-# Offense count: 66
+# Offense count: 64
 RSpec/MultipleExpectations:
   Max: 12
 
-# Offense count: 80
+# Offense count: 64
 # Configuration parameters: EnforcedStyle, IgnoreSharedExamples.
 # SupportedStyles: always, named_only
 RSpec/NamedSubject:
   Exclude:
     - 'spec/controllers/application_controller_spec.rb'
     - 'spec/controllers/file_controller_spec.rb'
-    - 'spec/controllers/iiif_controller_spec.rb'
-    - 'spec/controllers/iiif_token_controller_spec.rb'
     - 'spec/controllers/webauth_controller_spec.rb'
     - 'spec/models/approved_location_spec.rb'
     - 'spec/models/projection_spec.rb'
@@ -183,10 +177,10 @@ RSpec/NamedSubject:
     - 'spec/services/iiif_metadata_service_spec.rb'
     - 'spec/services/media_authentication_json_spec.rb'
 
-# Offense count: 54
+# Offense count: 47
 # Configuration parameters: AllowedGroups.
 RSpec/NestedGroups:
-  Max: 6
+  Max: 5
 
 # Offense count: 4
 RSpec/RepeatedDescription:
@@ -214,7 +208,6 @@ RSpec/VerifiedDoubles:
     - 'spec/models/purl_spec.rb'
     - 'spec/models/stacks_media_token_spec.rb'
     - 'spec/services/iiif_info_service_spec.rb'
-    - 'spec/services/media_authentication_json_spec.rb'
 
 # Offense count: 2
 # This cop supports safe autocorrection (--autocorrect).
@@ -247,7 +240,7 @@ Style/HashAsLastArrayItem:
   Exclude:
     - 'app/controllers/object_controller.rb'
 
-# Offense count: 13
+# Offense count: 12
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: PreferredDelimiters.
 Style/PercentLiteralDelimiters:
