+++ app/views/iiif/auth/v1/token/create.html.erb
