+++ spec/requests/iiif/auth/v1/tokens_spec.rb
@@ -2,66 +2,61 @@
 
 require 'rails_helper'
 
-RSpec.describe IiifTokenController do
-  render_views
-
+RSpec.describe 'IIIF auth v1 tokens' do
   let(:user) { User.new(anonymous_locatable_user: true) }
 
   before do
-    allow(controller).to receive(:current_user).and_return(user)
+    allow_any_instance_of(Iiif::Auth::V1::TokenController).to receive(:current_user).and_return(user)
   end
 
   describe '#create' do
-    context 'browser-based interaction' do
+    context 'when messageId is sent (browser-based interaction)' do
       let(:user) { User.new id: 'xyz', webauth_user: true }
 
-      subject do
-        get :create, params: { origin: 'http://example.edu/', messageId: '1' }
+      before do
+        get '/image/iiif/token?origin=http://example.edu/&messageId=1'
       end
 
       it 'wipes out the X-Frame-Options header' do
-        expect(subject.headers['X-Frame-Options']).to eq ''
+        expect(response.headers['X-Frame-Options']).to eq ''
       end
 
       it 'assigns the message and origin parameters' do
-        expect(subject.response_code).to eq 200
+        expect(response.response_code).to eq 200
         expect(assigns(:origin)).to eq 'http://example.edu/'
         expect(assigns(:message)).to include messageId: '1', accessToken: be_present
       end
 
-      context 'other formats' do
-        subject do
-          get :create, params: { origin: 'http://example.edu/', messageId: '1', format: :js }
+      context 'when the format is js' do
+        before do
+          get '/image/iiif/token.js?origin=http://example.edu/&messageId=1'
         end
 
         it 'renders HTML anyway' do
-          expect(subject.body).to match(/<html/)
+          expect(response.body).to match(/<html/)
         end
       end
 
-      context 'missing the origin header' do
-        subject do
-          get :create, params: { messageId: '1' }
-        end
-
+      context 'when the origin parameter is missing' do
         it 'returns a 400 error' do
-          expect { subject.response }.to raise_error ActionController::ParameterMissing
+          get '/image/iiif/token?messageId=1'
+          expect(response).to have_http_status(:bad_request)
         end
       end
     end
 
-    context 'JSON API interaction' do
-      subject do
-        get :create
-      end
-
+    describe 'JSON API interaction' do
       context 'with a user' do
+        before do
+          get '/image/iiif/token'
+        end
+
         let(:user) { User.new id: 'xyz', webauth_user: true }
 
         it 'returns the token response' do
-          expect(subject.status).to eq 200
+          expect(response.status).to eq 200
 
-          data = JSON.parse(subject.body)
+          data = response.parsed_body
           expect(data['accessToken']).not_to be_blank
           expect(data['tokenType']).to eq 'Bearer'
           expect(data['expiresIn']).to be > 0
@@ -69,11 +64,14 @@ RSpec.describe IiifTokenController do
       end
 
       context 'with an anonymous user' do
-        it 'returns the error response' do
-          expect(subject.status).to eq 401
+        before do
+          get '/image/iiif/token'
+        end
 
-          data = JSON.parse(subject.body)
-          expect(data['error']).to eq 'missingCredentials'
+        it 'returns the error response' do
+          expect(response.status).to eq 401
+
+          expect(response.parsed_body['error']).to eq 'missingCredentials'
         end
       end
     end
@@ -81,12 +79,11 @@ RSpec.describe IiifTokenController do
 
   describe '#create_for_item' do
     it 'returns an error response' do
-      get :create_for_item, params: { id: 'whatever', format: :js }
+      get '/image/iiif/token/whatever'
 
       expect(response.status).to eq 401
 
-      data = JSON.parse(response.body)
-      expect(data['error']).to eq 'missingCredentials'
+      expect(response.parsed_body['error']).to eq 'missingCredentials'
     end
 
     context 'with a token for the item' do
@@ -106,9 +103,8 @@ RSpec.describe IiifTokenController do
       end
 
       it 'calls the ordinary create method' do
-        allow(controller).to receive(:create)
-        get :create_for_item, params: { id: 'whatever', format: :js }
-        expect(controller).to have_received(:create)
+        get '/image/iiif/token/whatever'
+        expect(response.parsed_body).to include({ 'tokenType' => 'Bearer' })
       end
     end
   end
